+++ src/app/home/buscar-tutores/BuscarTutores.css
@@ -758,6 +758,40 @@
   margin-bottom: 0;
 }
 
+.course-availability-shell__tutor-profile {
+  margin-top: var(--space-4);
+}
+
+.course-availability-shell__main-body--detail {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
+
+.course-availability-shell__calendar-card {
+  min-width: 0;
+}
+
+.course-availability-shell__state {
+  display: flex;
+  min-height: 220px;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: var(--space-3);
+  padding: var(--space-6);
+  text-align: center;
+  color: var(--calico-body-muted);
+  background: #fff;
+  border: 1px solid var(--page-section-border);
+  border-radius: var(--page-section-radius);
+}
+
+.course-availability-shell__state--error {
+  color: #b42318;
+  border-color: rgba(180, 35, 24, 0.18);
+}
+
 .course-availability-shell__panel {
   min-width: 0;
   min-height: 0;
